Support devices on linux
